// apps/server/src/modules/common-cartridge/common-cartridge-client/room-client/room-api-client/models/board-column-board-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BoardColumnBoardResponse
 */
export interface BoardColumnBoardResponse {
    /**
     * 
     * @type {string}
     * @memberof BoardColumnBoardResponse
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof BoardColumnBoardResponse
     */
    'title': string;
    /**
     * 
     * @type {boolean}
     * @memberof BoardColumnBoardResponse
     */
    'published': boolean;
    /**
     * 
     * @type {string}
     * @memberof BoardColumnBoardResponse
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof BoardColumnBoardResponse
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof BoardColumnBoardResponse
     */
    'columnBoardId': string;
    /**
     * 
     * @type {string}
     * @memberof BoardColumnBoardResponse
     */
    'layout': string;
}